user bubble fixed
