+++ app/layout.tsx
@@ -3,6 +3,7 @@ import type { AppProps } from 'next/app'
 import { NextUIProvider } from "@nextui-org/react";
 import NavBar from './components/NavBar';
 import { FaUserCircle } from 'react-icons/fa';
+import UserBubble from "./components/UserBubble";
 
 import { Providers } from "./providers";
 
@@ -11,7 +12,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
     <html lang="en">
       <body>
         <Providers>
-          <div className="m-3 flex justify-end"><FaUserCircle size={100} /></div>
+        <div className="m-3 flex justify-end"><UserBubble /></div>
           <div className='flex'>
             <div className='z-10'><NavBar /></div>
             <div className='w-4/5 z-20'>{children}</div>
